聊天登录全量更新部分实现

+++ internal/common/res/chat.go
@@ -13,7 +13,7 @@ type GetMessageListRespond struct {
 	FileName   string `json:"fileName"`
 	FileType   string `json:"fileType"`
 	CreatedAt  string `json:"createdAt"`
-	MessageType int8 `json:"messageType"`
+	MessageType int32 `json:"messageType"`
 }
 
 // GetGroupMessageListRespond 群消息列表响应结构
